Nighttime
Took the daylight.php and let ChatGPT inverse it so that this nighttime.php now shows an event that has the length of the time between sunset on one day and sunrise the next day. I am not a developer so I did not check the code by ChatGPT but I did upload the file to my own server and it works. Someone might do some code review though as there might also be some redundant leftovers from copying from the daylight.php file. Cheers and enjoy.